+++ src/star-search/dtos/bing-search-results.dto.ts
@@ -1,0 +1,7 @@
+export class BingSearchResultDto {
+  title: string;
+  url: string;
+  displayUrl: string;
+  snippet: string;
+  datePublished: string;
+}