+++ src/timescale/pull_request_github_events.service.ts
@@ -1,6 +1,7 @@
 import { BadRequestException, Inject, Injectable, NotFoundException, forwardRef } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository, SelectQueryBuilder } from "typeorm";
+import { UserPrsDto } from "../user/dtos/user-prs.dto";
 import { ContributorHistogramDto } from "../histogram/dtos/contributor.dto";
 import { DbContributorHistogram } from "../histogram/entities/contributors.entity";
 import { PullRequestHistogramDto } from "../histogram/dtos/pull_request.dto";
@@ -44,7 +45,8 @@ export class PullRequestGithubEventsService {
     private pullRequestGithubEventsRepository: Repository<DbPullRequestGitHubEvents>,
     @Inject(forwardRef(() => RepoService))
     private readonly repoService: RepoService,
-    private readonly userListService: UserListService
+    @Inject(forwardRef(() => UserListService))
+    private userListService: UserListService
   ) {}
 
   baseQueryBuilder() {
@@ -198,7 +200,8 @@ export class PullRequestGithubEventsService {
   async getOpenedPrsCountForAuthor(
     username: string,
     contribType: ContributorStatsTypeEnum,
-    range: number
+    range: number,
+    repos?: string[]
   ): Promise<number> {
     const queryBuilder = this.pullRequestGithubEventsRepository.manager
       .createQueryBuilder()
@@ -207,6 +210,10 @@ export class PullRequestGithubEventsService {
       .where(`LOWER(actor_login) = '${username}'`)
       .andWhere("pr_action = 'opened'")
       .groupBy("LOWER(actor_login)");
+
+    if (repos && repos.length > 0) {
+      queryBuilder.andWhere(`LOWER(repo_name) IN (:...repos)`, { repos });
+    }
 
     applyContribTypeEnumFilters(contribType, queryBuilder, range);
 
@@ -286,19 +293,28 @@ export class PullRequestGithubEventsService {
     return countResult.count !== 0;
   }
 
-  async getOpenedPullReqEventsForLogin(username: string, range: number): Promise<DbPullRequestGitHubEvents[]> {
+  async getOpenedPullReqEventsForLogin(
+    username: string,
+    range: number,
+    repos?: string[]
+  ): Promise<DbPullRequestGitHubEvents[]> {
     const queryBuilder = this.baseQueryBuilder()
       .where(`LOWER(actor_login) = '${username}'`)
       .andWhere("pr_action = 'opened'")
       .andWhere(`event_time > NOW() - INTERVAL '${range} days'`);
 
+    if (repos && repos.length > 0) {
+      queryBuilder.andWhere(`LOWER(repo_name) IN (:...repos)`, { repos });
+    }
+
     return queryBuilder.getMany();
   }
 
-  async findAllByPrAuthor(author: string, pageOptionsDto: PageOptionsDto): Promise<PageDto<DbPullRequestGitHubEvents>> {
+  async findAllByPrAuthor(author: string, pageOptionsDto: UserPrsDto): Promise<PageDto<DbPullRequestGitHubEvents>> {
     const startDate = GetPrevDateISOString(pageOptionsDto.prev_days_start_date);
     const range = pageOptionsDto.range!;
     const order = pageOptionsDto.orderDirection!;
+    const repos = pageOptionsDto.repos ? pageOptionsDto.repos.toLowerCase().split(",") : undefined;
 
     /*
      * because PR events may be "opened" or "closed" many times, this inner CTE query gets similar PRs rows
@@ -314,6 +330,10 @@ export class PullRequestGithubEventsService {
       .andWhere(`'${startDate}'::TIMESTAMP >= "pull_request_github_events"."event_time"`)
       .andWhere(`'${startDate}'::TIMESTAMP - INTERVAL '${range} days' <= "pull_request_github_events"."event_time"`)
       .orderBy("event_time", order);
+
+    if (repos && repos.length > 0) {
+      cteBuilder.andWhere(`LOWER("pull_request_github_events"."repo_name") IN (:...repos)`, { repos });
+    }
 
     return this.execCommonTableExpression(pageOptionsDto, cteBuilder);
   }